--- src/features/bookmark/components/sections/BookmarkList.tsx
@@ -1,16 +1,0 @@
-import { Bookmark } from '@/features/bookmark/types';
-import BookmarkListItem from '@/features/bookmark/components/sections/BookmarkListItem';
-
-interface BookmarkListProps {
-  bookmarks: Bookmark[];
-}
-
-export default function BookmarkList({ bookmarks }: BookmarkListProps) {
-  return (
-    <div className="flex flex-col gap-3">
-      {bookmarks.map((bookmark) => (
-        <BookmarkListItem key={bookmark.postId} bookmark={bookmark} />
-      ))}
-    </div>
-  );
-}
