+++ src/features/bookmark/components/sections/BookmarkSection.tsx
@@ -1,11 +1,11 @@
 'use client';
 
-import BookmarkEmpty from '@/features/bookmark/components/ui/BookmarkEmpty';
-import BookmarkList from '@/features/bookmark/components/ui/BookmarkList';
-import { useBookmarks } from '@/features/bookmark/queries';
+import MyBookmarkEmpty from '@/features/bookmark/components/ui/MyBookmarkEmpty';
+import MyBookmarkList from '@/features/bookmark/components/ui/MyBookmarkList';
+import { useGetBookmarks } from '@/features/bookmark/queries';
 
 export default function BookmarkSection() {
-  const { data: bookmarks, isLoading, isError, error } = useBookmarks();
+  const { data: bookmarks, isLoading, isError, error } = useGetBookmarks();
 
   return (
     <>
@@ -17,7 +17,7 @@ export default function BookmarkSection() {
       </div>
 
       {/* 빈 리스트 or 북마크 리스트 */}
-      {!bookmarks || bookmarks?.length === 0 ? <BookmarkEmpty /> : <BookmarkList />}
+      {!bookmarks || bookmarks?.length === 0 ? <MyBookmarkEmpty /> : <MyBookmarkList />}
     </>
   );
 }
