+++ src/features/bookmark/components/ui/MyBookmarkList.tsx
@@ -1,0 +1,15 @@
+'use client';
+import MyBookmarkListItem from '@/features/bookmark/components/ui/MyBookmarkListItem';
+import { useGetBookmarks } from '@/features/bookmark/queries';
+
+export default function BookmarkList() {
+  const { data: bookmarks } = useGetBookmarks();
+
+  return (
+    <div className="flex flex-col gap-3">
+      {bookmarks?.map((bookmark) => (
+        <MyBookmarkListItem key={bookmark.postId} bookmark={bookmark} />
+      ))}
+    </div>
+  );
+}
